Refactor CSVLoaderService into separate class
Split CSVLoaderService into its own file for better modularity
and maintainability. Removed the duplicate implementation
from CSVLoaderNewSearch.
